+++ src/layer/background-layer-group.ts
@@ -1,8 +1,7 @@
 import OlMap from 'ol/Map.js';
-import { LayerGroup, type LayerGroupOptions } from './layer-group.js';
-import { LayerUidKey } from './property-key.js';
+import { LayerGroup, type LayerGroupOptions, LayerUidKey } from './layer-group.js';
 
-export const DefaultLayerBGGroupName = 'olcBackgroundLayerGroup';
+export const DefaultLayerBGGroupUid = 'olcBackgroundLayerGroupUid';
 
 /**
  * LayerGroup specialized to manage background layers (mostly tiled layers).
@@ -11,7 +10,7 @@ export const DefaultLayerBGGroupName = 'olcBackgroundLayerGroup';
  */
 export class BackgroundLayerGroup extends LayerGroup {
   constructor(map: OlMap, options: LayerGroupOptions = {}) {
-    const layerGroupUid = options[LayerUidKey] || DefaultLayerBGGroupName;
+    const layerGroupUid = options[LayerUidKey] || DefaultLayerBGGroupUid;
     super(map);
     const position = options.position ?? 0;
     this.addLayerGroup(layerGroupUid, position);