+++ CHANGES.md
@@ -3,8 +3,9 @@
 List changes at major/minor level. Patches should be straightforward.
 
 ## 1.1.0
+- Rework interactions to have generic groups to manage interactions.
 - Add features in the event fired by the "setFeatureProperty" method. Add more flexibility.
-- Replace "selectFeatures" with a more generical "featureAffected" method and event.
+- Replace "selectFeatures" with a more generic "featureAffected" method and event.
 - Remove layerAdded observable; Use native collection events instead.
 - Move EmptyStyle to a dedicated `src/style.ts` file.
 - Remove CommonProperties from the `src/layer/layer-group.ts` file and move the `LayerUid` key into